:hammer: Rework architecture.
